# PERF/BUG: improve factorize for datetimetz