@jreback 2016-07-22
lgtm.

I *think* this may solve some of these Ambiguous time issues: https://github.com/pydata/pandas/search?q=cannot+infer+dst&type=Issues&utf8=%E2%9C%93

maybe