@jreback 2016-07-24
thanks! see my comment above if you have some time.