@sinhrks 2016-07-24
yeah, i think we can remove most of `tz_localize(None)` logic from core.